# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "description": "Api for Ecommerce Application",
    "title": "Ecommerce Api"
  },
  "servers": [
    {
      "url": "http://localhost:3200"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "header",
        "name": "authorization",
        "type": "apiKey"
      }
    }
  },
  "paths": {
    "/api/products": {
      "get": {
        "tags": ["products"],
        "summary": "Get Products",
        "description": "User will get all products",
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/carts": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get Cart Items",
        "description": "Retrieve items in the user's cart.",
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CartItem"
                  }
                }
              }
            }
          }
        }
      }
    },
      "/api/carts/": {
          "post": {
              "tags": ["Cart"],
              "summary": "Add Item to Cart",
              "description": "Add an item to the user's cart.",
              "security": [
                {
                  "JWT": {}
                }
              ],
              "parameters": [
                  {
                      "in": "query",
                      "name": "productId",
                      "description": "ID of the product to add to the cart",
                      "schema": {
                          "type": "integer"
                      },
                      "required": true
                  },
                  {
                      "in": "query",
                      "name": "quantity",
                      "description": "Quantity of the product to add to the cart",
                      "schema": {
                          "type": "integer"
                      },
                      "required": true
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful operation",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "object",
                                  "properties": {
                                      "res": {
                                          "type": "object"
                                      },
                                      "msg": {
                                          "type": "string"
                                      }
                                  }
                              }
                          }
                      }
                  },
                  "400": {
                      "description": "Bad request, invalid parameters",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "type": "object"
                              }
                          }
                      }
                  }
              }
          }
      },
...
  }
}